+++ config/dev.exs
@@ -1,13 +1,5 @@
 import Config
 
-config :payroll_api, PayrollApi.Repo,
-  username: System.get_env("POSTGRES_USER", System.get_env("USER")),
-  password: System.get_env("POSTGRES_PASSWORD"),
-  hostname: System.get_env("POSTGRES_HOST", "localhost"),
-  database: System.get_env("POSTGRES_DB", "payroll_api_dev"),
-  show_sensitive_data_on_connection_error: true,
-  pool_size: 10
-
 # For development, we disable any cache and enable
 # debugging and code reloading.
 #