+++ lib/payroll_api/repo.ex
@@ -1,5 +1,0 @@
-defmodule PayrollApi.Repo do
-  use Ecto.Repo,
-    otp_app: :payroll_api,
-    adapter: Ecto.Adapters.Postgres
-end