+++ test/repo_test.exs
@@ -1,7 +1,0 @@
-defmodule PayrollApi.RepoTest do
-  use ExUnit.Case, async: true
-
-  test "queries PostgreSQL" do
-    assert %{rows: [[1]]} = PayrollApi.Repo.query!("SELECT 1")
-  end
-end